+++ src/libgcm/mesh/cube/BasicCubicMesh.cpp
@@ -55,12 +55,12 @@ BasicCubicMesh::~BasicCubicMesh()
 void BasicCubicMesh::preProcessGeometry()
 {
     LOG_DEBUG("Preprocessing mesh geometry started.");
-	for(int i = 0; i < getNodesNumber(); i++)
+	for(uint i = 0; i < getNodesNumber(); i++)
 		getNodeByLocalIndex(i).setBorder(false);
-    for(int i = 0; i < getNodesNumber(); i++)
+    for(uint i = 0; i < getNodesNumber(); i++)
     {
         Node& node = getNodeByLocalIndex(i);
-        for( int i = 0; i < 3; i ++)
+        for( uint i = 0; i < 3; i ++)
         {
             if( ( fabs(node.coords[i] - outline.min_coords[i]) < EQUALITY_TOLERANCE )
                 || ( fabs(node.coords[i] - outline.max_coords[i]) < EQUALITY_TOLERANCE ) )
@@ -110,7 +110,7 @@ void BasicCubicMesh::calcMinH()
     float h;
 
     // We suppose that mesh is uniform
-    for(int i = 1; i < getNodesNumber(); i++)
+    for(uint i = 1; i < getNodesNumber(); i++)
     {
         Node& node = getNodeByLocalIndex(i);
         h = distance(base.coords, node.coords);
@@ -148,7 +148,7 @@ void BasicCubicMesh::findBorderNodeNormal(const Node& node, float* x, float* y, 
     normal[0] = normal[1] = normal[2] = 0.0;
 
 	uint i = node.contactDirection;
-	for( int cntr = 0; cntr < 3; cntr ++) {
+	for( uint cntr = 0; cntr < 3; cntr ++) {
         if( fabs(node.coords[i] - outline.min_coords[i]) < EQUALITY_TOLERANCE ) {
             normal[i] = -1;
             break;
@@ -164,12 +164,12 @@ void BasicCubicMesh::findBorderNodeNormal(const Node& node, float* x, float* y, 
     *z = normal[2];
 };
 
-int BasicCubicMesh::findNeighbourPoint(Node& node, float dx, float dy, float dz,
+uint BasicCubicMesh::findNeighbourPoint(Node& node, float dx, float dy, float dz,
 	bool debug, float* coords, bool* innerPoint)
 {
-    //int meshSizeX = 1 + (outline.maxX - outline.minX + meshH * 0.1) / meshH;
-	int meshSizeY = 1 + (outline.maxY - outline.minY + meshH * 0.1) / meshH;
-	int meshSizeZ = 1 + (outline.maxZ - outline.minZ + meshH * 0.1) / meshH;
+    //uint meshSizeX = 1 + (outline.maxX - outline.minX + meshH * 0.1) / meshH;
+	uint meshSizeY = 1 + (outline.maxY - outline.minY + meshH * 0.1) / meshH;
+	uint meshSizeZ = 1 + (outline.maxZ - outline.minZ + meshH * 0.1) / meshH;
 
     assert_le(vectorSquareNorm(dx, dy, dz), getMinH() * getMinH() * (1 + EQUALITY_TOLERANCE) );
 
@@ -187,7 +187,7 @@ int BasicCubicMesh::findNeighbourPoint(Node& node, float dx, float dy, float dz,
         return -1;
     }
 
-    int neighNum = node.number;
+    uint neighNum = node.number;
     if ( dx > EQUALITY_TOLERANCE )
         neighNum += meshSizeY*meshSizeZ;
     else if ( dx < -EQUALITY_TOLERANCE )
@@ -215,26 +215,26 @@ bool BasicCubicMesh::interpolateNode(Node& origin, float dx, float dy, float dz,
 		return true;
 	}
 
-    int neighInd = findNeighbourPoint( origin, dx, dy, dz, debug,
+    uint neighInd = findNeighbourPoint( origin, dx, dy, dz, debug,
                                     targetNode.coords, &isInnerPoint );
 
-    if( neighInd == -1 )
+    if( neighInd == (uint) -1 )
         return false;
 
     Node tmpNode;
-    int secondNeighInd = findNeighbourPoint( origin, -dx, -dy, -dz, debug,
+    uint secondNeighInd = findNeighbourPoint( origin, -dx, -dy, -dz, debug,
                                     tmpNode.coords, &isInnerPoint );
 
     //interpolator1->interpolate( targetNode, origin, getNode( neighInd ) );
     //return true;
     
-    if( secondNeighInd == -1 )
+    if( secondNeighInd == (uint) -1 )
     {
         interpolator1->interpolate( targetNode, origin, getNode( neighInd ) );
     }
     else
     {
-        int leftInd, rightInd;
+        uint leftInd, rightInd;
         if( dx + dy + dz > 0 )
         {
             leftInd = secondNeighInd;
@@ -259,7 +259,7 @@ bool BasicCubicMesh::interpolateNode(Node& node)
 bool BasicCubicMesh::interpolateBorderNode_old(real x, real y, real z,
                                 real dx, real dy, real dz, Node& node)
 {
-    //int meshSizeX = 1 + (outline.maxX - outline.minX + meshH * 0.1) / meshH;
+    //uint meshSizeX = 1 + (outline.maxX - outline.minX + meshH * 0.1) / meshH;
     float coords[3];
     float tx = coords[0] = x + dx;
     float ty = coords[1] = y + dy;
@@ -269,8 +269,8 @@ bool BasicCubicMesh::interpolateBorderNode_old(real x, real y, real z,
     {
         // FIXME_ASAP
         float minH = std::numeric_limits<float>::infinity();
-        int num = -1;
-        for(int i = 1; i < getNodesNumber(); i++) {
+        uint num = -1;
+        for(uint i = 1; i < getNodesNumber(); i++) {
             Node& node = getNodeByLocalIndex(i);
 			if(node.isBorder()) {
 				float h = distance(coords, node.coords);
@@ -288,21 +288,21 @@ bool BasicCubicMesh::interpolateBorderNode_old(real x, real y, real z,
     return false;
 };
 
-void BasicCubicMesh::findNearestsNodes(const vector3r& coords, int N, vector< pair<int,float> >& result)
+void BasicCubicMesh::findNearestsNodes(const vector3r& coords, uint N, vector< pair<uint,float> >& result)
 {
 	int n = 0;//floor( pow( (float)(N), 1.0 / 3.0 ) );
 
-	int i_min =	max( int( (coords[0] - outline.minX) / meshH ) - n, 0);
-	int i_max =	min( int( (coords[0] - outline.minX) / meshH ) + 1 + n, numX);
-	int j_min =	max( int( (coords[1] - outline.minY) / meshH ) - n, 0);
-	int j_max =	min( int( (coords[1] - outline.minY) / meshH ) + 1 + n, numY);
-	int k_min =	max( int( (coords[2] - outline.minZ) / meshH ) - n, 0);
-	int k_max =	min( int( (coords[2] - outline.minZ) / meshH ) + 1 + n, numZ);
+	uint i_min =	max( int( (coords[0] - outline.minX) / meshH ) - n, 0);
+	uint i_max =	min( int( (coords[0] - outline.minX) / meshH ) + 1 + n, (int) numX);
+	uint j_min =	max( int( (coords[1] - outline.minY) / meshH ) - n, 0);
+	uint j_max =	min( int( (coords[1] - outline.minY) / meshH ) + 1 + n, (int) numY);
+	uint k_min =	max( int( (coords[2] - outline.minZ) / meshH ) - n, 0);
+	uint k_max =	min( int( (coords[2] - outline.minZ) / meshH ) + 1 + n, (int) numZ);
 
-	int num;
-	for( int k = k_min; k <= k_max; k++ )
-		for( int j = j_min; j <= j_max; j++ )
-			for( int i = i_min; i <= i_max; i++ )
+	uint num;
+	for( uint k = k_min; k <= k_max; k++ )
+		for( uint j = j_min; j <= j_max; j++ )
+			for( uint i = i_min; i <= i_max; i++ )
 	        {
 				num = i * (numY + 1) * (numZ + 1) + j * (numZ + 1) + k + nodes[0].number;
 				Node& node = getNode(num);
@@ -322,7 +322,7 @@ bool BasicCubicMesh::interpolateBorderNode(real x, real y, real z,
 	
 	if( !outline.isInAABB(x, y, z) )
 	{
-		vector< pair<int,float> > result;
+		vector< pair<uint,float> > result;
 
 		findNearestsNodes(coords, N, result);
 
@@ -342,32 +342,32 @@ bool BasicCubicMesh::interpolateBorderNode(real x, real y, real z,
     return false;
 };
 
-void BasicCubicMesh::setNumX(int _numX)
+void BasicCubicMesh::setNumX(uint _numX)
 {
 	numX = _numX;
 }
 
-int BasicCubicMesh::getNumX() const
+uint BasicCubicMesh::getNumX() const
 {
 	return numX;
 }
 
-void BasicCubicMesh::setNumY(int _numY)
+void BasicCubicMesh::setNumY(uint _numY)
 {
 	numY = _numY;
 }
 
-int BasicCubicMesh::getNumY() const
+uint BasicCubicMesh::getNumY() const
 {
 	return numY;
 }
 
-void BasicCubicMesh::setNumZ(int _numZ)
+void BasicCubicMesh::setNumZ(uint _numZ)
 {
 	numZ = _numZ;
 }
 
-int BasicCubicMesh::getNumZ() const
+uint BasicCubicMesh::getNumZ() const
 {
 	return numZ;
 }
